Move githubmentions to /internal/collector
